Added 4/17/2019
Added a test I2C file using better practice. Added Test BQ27441 Code. Changed data types for all files to  typedef data types.

+++ Delay.c
@@ -91,8 +91,8 @@ void Null_Timer_1 (void){
  * 
  * Returns: NULL (VOID).
  ******************************************************************************/
-void Delay_ms(int ms){
-    int i;
+void Delay_ms(uint16_t ms){
+    uint16_t i;
     for(i = 0; i <= ms;i++){
         IFS0bits.T1IF = 0;
         Timer_1_Reset();
