+++ Delay.h
@@ -25,6 +25,6 @@ void Timer_1_Stop(void);
 void Timer_1_Reset(void);
 void Null_Timer_1 (void);
 int Timer_1_Read(void);
-void Delay_ms(int);
+void Delay_ms(uint16_t);
 
 #endif
